feature(refactor)
Now using models, config and repo from DBManager
