+++ README.md
@@ -6,13 +6,14 @@ These instructions will get you a copy of the project up and running on your loc
 
 ## Prerequisites
 
-1. Install Java and set JAVA_HOME and PATH variables.
-2. Install Maven.
-3. Install Elasticsearch 5.6.0
-4. Configure ElasticSearch Cluster by opening ${ELASTICSEARCH_HOME}\config\elasticsearch.yml and adding the following configuration `cluster.name: pocket-cluster`
-5. Start Elasticsearch Instance
+For now, you need to have the DBManager clone on your local machine. 
+Follow all the steps that are in the DBManager's readme, except that you don't have to run it (**but you still have to build it** using `mvn clean install`)
 
 ### Running
+
+1. Start Elasticsearch Instance
+
+*(Go to project root)*
 
 `mvn clean install`
 `mvn spring-boot:run`
