--- app/src/main/java/com/jxc/app/exceptions/UserNotFoundException.java
@@ -1,12 +1,0 @@
-package com.jxc.app.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
-
