--- app/src/main/java/com/jxc/app/models/Address.java
@@ -1,15 +1,0 @@
-package com.jxc.app.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Address {
-    private String postalCode;
-    private String street;
-    private int streetNumber;
-    private String complement;
-}
