--- app/src/main/java/com/jxc/app/models/UserInfosDTO.java
@@ -1,18 +1,0 @@
-package com.jxc.app.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.elasticsearch.annotations.Field;
-import org.springframework.data.elasticsearch.annotations.FieldType;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserInfosDTO {
-
-    private String email;
-    private String username;
-    private Address address;
-}
