--- app/src/main/java/com/jxc/app/repositories/UserRepository.java
@@ -1,13 +1,0 @@
-package com.jxc.app.repositories;
-
-import com.jxc.app.models.User;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends ElasticsearchRepository<User, String> {
-
-    Page<User> findAll();
-}
