--- app/src/main/resources/application.properties
@@ -1,9 +1,0 @@
-server.port=8082
-
-elasticsearch.clustername = pocket-cluster
-elasticsearch.host = localhost
-elasticsearch.port = 9300
-elasticsearch.home = D:/Programmes/elasticsearch/5.6.0
-
-url.version = "/api/v1"
-url.users = "/users"
