--- app/src/test/java/com/jxc/app/AppApplicationTests.java
@@ -1,16 +1,0 @@
-package com.jxc.app;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class AppApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
