+++ src/main/java/com/jxc/readapis/exceptions/UserNotFoundException.java
@@ -1,0 +1,11 @@
+package com.jxc.readapis.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
