+++ src/test/java/com/jxc/readapis/ReadapisApplicationTests.java
@@ -1,0 +1,17 @@
+package com.jxc.readapis;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ReadapisApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
